+++ NEWS.md
@@ -32,3 +32,7 @@
 ### Bug Fixes
 
 - `show_typst(io, ::AbstractString)` now correctly escapes double quotation marks and dollar signs.
+- `show_typst(io, ::Complex)`
+    - Fix `Complex{Bool}`
+    - Handle negative imaginary part
+    - Passes component parts to `show_typst`